Added protection from null/undefined actor parameter to "setOpponentLive" method
I found that when you wait for the first cutscene and credits to pass, the game tries to start the "demo" level 0. But it has no guards[1] defined, so the it crashes leaving this error on the console: 
"phaser.js:74099 TypeError: Cannot read property 'charName' of undefined(…)xhr.onload @ phaser.js:74099"
I added a protection for this case.